fix(kurtosis-devnet): forbid tests from using local kurtosis context
